+++ src/api/routes/store/live-rates-post.js
@@ -1,14 +1,22 @@
-import { Validator } from "medusa-core-utils"
-import { shippoAddress, shippoLineItems, shippoRates } from "../../../utils/shippo"
+import {
+  shippoAddress,
+  shippoLineItem,
+  shippoRates,
+} from "../../../utils/shippo"
+import { validateShippingAddress } from "../../../utils/validator"
 
 export default async (req, res) => {
   const { cart_id } = req.body
   const cartService = req.scope.resolve("cartService")
   const totalsService = req.scope.resolve("totalsService")
   const shippingProfileService = req.scope.resolve("shippingProfileService")
-  const customShippingOptionService = req.scope.resolve("customShippingOptionService")
+  const customShippingOptionService = req.scope.resolve(
+    "customShippingOptionService"
+  )
   const manager = req.scope.resolve("manager")
-  const customShippingOptionRepository = req.scope.resolve("customShippingOptionRepository")
+  const customShippingOptionRepository = req.scope.resolve(
+    "customShippingOptionRepository"
+  )
 
   const cart = await cartService.retrieve(cart_id, {
     relations: [
@@ -18,12 +26,12 @@ export default async (req, res) => {
       "items.variant",
       "items.variant.product",
       "discounts",
-      "region"
-    ]
+      "region",
+    ],
   })
 
   // Validate if cart has a complete shipping address
-  const validAddress = Validator.shippingAddress().validate(cart.shipping_address)
+  const validAddress = validateShippingAddress(cart.shipping_address)
   if (validAddress.error) {
     return next(
       new MedusaError(
@@ -34,40 +42,51 @@ export default async (req, res) => {
   }
 
   const shippingOptions = await shippingProfileService.fetchCartOptions(cart)
-  const lineItems = await shippoLineItems(cart, totalsService)
+
+  const lineItems = await Promise.all(
+    cart.items.map(async (item) => {
+      const totals = await totalsService.getLineItemTotals(item, cart)
+      return shippoLineItem(item, totals.subtotal, cart.region.currency_code)
+    })
+  )
+
   const toAddress = shippoAddress(cart.shipping_address, cart.email)
   const rates = await shippoRates(toAddress, lineItems, shippingOptions)
 
-  const customShippingOptions = await customShippingOptionService.list({ cart_id: cart_id })
-    .then(async cartCustomShippingOptions => {
-
+  const customShippingOptions = await customShippingOptionService
+    .list({ cart_id })
+    .then(async (cartCustomShippingOptions) => {
       if (cartCustomShippingOptions.length) {
         const customShippingOptionRepo = await manager.getCustomRepository(
           customShippingOptionRepository
         )
 
         await customShippingOptionRepo.remove(cartCustomShippingOptions)
-
       }
+
       return await Promise.all(
-        shippingOptions.map(async option => {
+        shippingOptions.map(async (option) => {
           const optionRate = rates.find(
-            rate => rate.title == option.data.name
+            (rate) => rate.title == option.data.name
           )
 
-          const price = (optionRate.amount_local) || optionRate.amount
+          const price = optionRate.amount_local || optionRate.amount
 
-          return await customShippingOptionService.create({
-            cart_id: cart_id,
-            shipping_option_id: option.id,
-            price: parseInt(parseFloat(price) * 100)
-          },{
-            metadata: {
-              is_shippo_rate: true,
-              ...optionRate
+          return await customShippingOptionService.create(
+            {
+              cart_id,
+              shipping_option_id: option.id,
+              price: parseInt(parseFloat(price) * 100),
+            },
+            {
+              metadata: {
+                is_shippo_rate: true,
+                ...optionRate,
+              },
             }
-          })
-        }))
+          )
+        })
+      )
     })
 
   res.json({ customShippingOptions })